requirements.txt
matplotlib>=3.3.4
numpy>=1.19.5
opencv-python>=4.5
pandas>=1.1
Pillow>=8.4
scipy>=1.5
sqlalchemy>=1.4.39
flask>=2.1.3
flask_sqlalchemy>=2.5.1
torch>=1.10
torchvision>=0.11
# face_detection@git+https://github.com/elliottzheng/face-detection